requirements.txt
ultralytics
Pillow
python-dotenv
requests
